requirements.txt
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
googletrans==3.1.0a0
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==8.4.0
playsound==1.2.2
PyAudio @ file:///C:/Users/Evan/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pypiwin32==223
pyttsx3==2.90
pywin32==301
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SpeechRecognition==3.8.1
urllib3==1.26.6
